jump-or-exec -> mn-jamp-or-exec, root-menu splited, autostart added, and minor
A Damien Elmes' jump-or-exec has been replaced by mn-jump-or-exec which is
way shorter and written inside of the configuration file thus it is no
longer dependent on external files.  Also, the function is a bit less
complex so maybe it'll work few ms faster. :P

root-menu has been splited into apps-menu (bound to Button3-Click, contains
applications) and root-menu (bound to W-Button3-Click, contains what was
previously in Exit sub menu).  This way it is much harder to exit sawfish by
mistake. ;)

An autostart has been added, ie. a list of applications that need to be
executed after sawfish starts.  The code uses similar aproach as in
mn-jump-or-exec (expect for the part when existing window gets focus) so two
instances of the same program won't be executed after sawfish restarts.

As of other changes, focus-mode has been changed to 'enter-click so now I
can focus a window by clicking it if it ever happens that my pointers is
above a unfocused window (previously I had to mouve mouse out and in).

And last, a copyright info has added at the begining of the file.