Added code for detecting host and added host dependent configuration
Arrangement of windows is different on different hosts mostly due to a
different resolution.  Also, on some machines I don't run certain
applications and some bindings are useless.  Therefore the configuration
file has to detect what machine it is running on and turn some options
on or off depending on the machine.