[emacs] A few small changes
indent-or-complete now uses char-syntax with char-before and
char-after instead of looking-at.  Also, M-/ now calls
indent-or-complete-complete-function.

In XML/HTML mode, non breaking space is inserted as &#160;
instead of &nbsp;.  This is because named version need not
be present in XML DTDs and nXML highlights it as error.

And finally, MLDonkey support added.