[bin] New directory for executable
A new directory has been created for storing scripts that go into
the ~/bin directory.  term, which was previously kept in the root
of the repository, has been moved there and a new file -- s -- has
been added.

s is a sudo-what-I-mean and poor-man's-sudo at the same time.
When called with no arguments it acts as if "sudo -s" or "su"
was called (this do-what-I-mean capability) and when called
with arguments it acts as if "sudo" was invoked with specified
command even if sudo is not available (this poor-man's-sudo).