[sawfish] Updated bindings to support Dvorak
On dvorak some keys are in different locations, ;) and since most
of my bindings are based on locations of the keys and not the
letters they represent additional bindings are needed for Dvorak.

Problem is with Q which should be Play on Dvorak but because it's
Emacs on qwerty I left it as such.  As a side note, Win+C/Win+J
is now Toggle rather then Pause so I can resume the music even
without Play.