[bash] Changed g() to call ~/bin/git
The g() function has been changed so that instead of calling
git directly it calls ~/bin/g.  ~/bin/g has been added as a
symlink to /usr/bin/git so the change seems pointless but the
thing is, that on specific machines ~/bin/g can be replaced by
some git wrapper.