[emacs] ~/bin/e changes
This commit changes ~/bin/e so that calling it with no arguments brings
Emacs frame.  It currently creates a new frame even if one is already
visible on given display.

More over, ~/bin/e now assumes that emacsclient is present on
the system.

And finally, the usage where sourcing ~/bin/e would define a function
e is now removed.