[a] apt-get, apt-file support
This commit adds support for apt-get (if aptitude is missing
apt-get is used) as well as apt-file (which is called on
update thus “a update” updates packages repository as well
as apt-file cache).

This commit also adds an “upgrade” action which, even though
deprecated by aptitude, I use all the time.