[sawfish] Added logging to a log file in /tmp
This commit makes everything sawfish prints on standard
error go to log file in /tmp.  When running sawfish,
console is not visible anyways so it's not very useful
getting error messages on it.  Moreover, since I use
tmpfs on /tmp, the log file disappears with each reboot
which is perfect.