[sawfish] Make Win key behave
I had quite a few problems with hyper vs. super being a win
key.  In X configuration Win keys are translated to Super_*
but for some strange reason sawfish failed to always
recognise it as super and on some machines I had to set
modifier to be hyper.  This seems, however, to be resolved
and thus I'm updating configuration.