[sawfish] Make viewport manipulation bindigs using loop
This commit changes the way W-# and W-F# bindings are
created.  Instead of overly long list, a loop is now used.
This saves a few lines of code (even though it appears
less readable ;) ).  This change, however, make
 set-current-window-viewport-x macro obsolete since it's
no longer used over and over again, and instead it is
berried inside a loop.