[sawfish] Transparency fixes.
Most impartontly, this commit adds an after-add-window-hook
which handles windows that are already opened when sawfish
starts (most likely to happen when sawfish restarts).
Because this hook is only useful during initialisation,
an after-initialization-hook removes it.

Two other tiny changes this commit introduces are: (i) defvar
is with define (no need for dynamic scoping after all), and
(ii) fixed indention in window-alpha-property.