[sawfish] Refactor various commands bindings
This commit changes the way bindings which call the system
function are defined.  This includes a new binding for
locking screen (which will work only if xscreen-saver is
present), volume control bindings, and MPD bindings.

Instead of using helper macros or adding an overly long
lambda definitions, those bindings are now set using
a loop which creates the lambda functions as needed.

This commit also removes code that sets value of MPD_HOST
environmental variable.  This should be moved to shell
initialisation scripts or similar.