[sawfish] Rewrite application launching and menus
This commit changes the way applacations are launched and
bindings defined.

Most importantly, mn-jump-or-exec is removed in favour of
mn-exec-unless function.  The former matched window's name
which was error-prone (eg. if I had Emacswiki opened in
Opera it was treated as Emacs window), whereas the latter
matches window class which matches exactly what I need.

The mn-exec-unless also greatly simplified launching of
terminal windows, since it also supports matching windows
in current viewport only.

Second of all, mn-machine and case-equal functions has been
removed.  The former was used by after-initialization-hook
but now instead the hook uses value of system-name function
directly; the letter was used in a few places to create a
command with geometry but this has been dropped since the
plan is to handle geometry of all windows in window matcher.

Bindings are now defined with a loop which removes the need for
any helper macros and/or long lambda expressions.

Some bindings has been removed as well (most noticeably
qwerty-compatibility ones) since I've never used them anyway.

Finally, this commit also removes the use of system function.
Instead, whenever possible a start-process is used launching
directly the command we want.  When this is not enough
(see exit-sawfish function) "/bin/sh -c" is called with
start-process.  mn-exec-unless also has a "compatibility"
mode where command is called through "/bin/sh -c" but best
to my knowledge, it is never used.