[emacs, gnus] Add set-key macro.
This commit adds a set-key macro that acts as a DWIM macro for
defining bindings.  It can be used to create bindings in global,
local and specified key-map.  It automatically guesses whether
you are creating binding to a symbol or a lambda function without
the need to type lambda boilerplate.

This commit also converts several calls to add-hook by calls to
add-lambda-hook in the Gnus configuration file.