[sawfish] Execute commands using `system' plus some semi related changes
This commit replaces the use of `start-process' with calling `system'.
The reason to do that is because librep kills all running processes
as it terminates.  This means that on Sawfish restart, all
applications started from Sawfish are killed.

By using `system' I can consistently add an ampersand to the end of
each command so that a double fork with the middle process termination
is performed so that the command can continue even when librep
exits.

This actually simplifies things as there is no longer need to specify
commands as lists but rather as a single string.  The benefit of the
former were not used at all in the configuration file, so there's no
need to worry about thouse.

To make that change, `mn-exec-unless' function has been renamed to
`mn-run-unless' and a new `mn-run' function has been added which calls
`system' and runs command.

In the meantime fewer other changes has been performed:

A new W-Enter binding has been added to windows which pops window
operations menu.  This is nice since on my configuration of Kinesis
keyboard, W-Enter is easy to press with just the left hand.

The W-Button3-Click binding on the root window has ben changed to
Button3-Click.  The need to press win key was just a nuisance.

The root menu has been mistake-proofed.  Each option (other then
“Customize”) pops a sub-menu with a single “Yes” option.  This means
that pressing one of the top level options by mistake won't kill
Sawfish and everything else.

XF86Audio* bindings has been removed since I no longer use a keyboard
which has multimedia keys (and in fact never used them anyway).

The W-l binding runs "xlocker" rather then locking xscreensaver.  The
idea is for "xlocker" command be a script in ~/bin which calls xlock.
This change is because I've stopped using xscreensaver in favour of
xlock.