[emacs] Add "google" C style
This commit adds the "google" C style which is C++ style used at
Google.  It uses the google-c-stale package from the repositories.
Obviously, this commit also adds a bit of c-common-mode-hook to set
the style to "google" if the file has "/google3/" in its path.

This commit also changes the way the default "mina86" style is
defined and loaded by moving the definition to a c-initialization-hook
rather than using eval-after-load.