[emacs] Rewrite `just-one-space'
From user's point of view, this commit changes the M-SPC binding
so that it cycles through: only one space, no spaces and the
original spacing.  This way, it can be used to join two lines
together to see if they fit desired column limit, and if not,
get the original spacing.

This is done, by practically rewriting the `just-one-space' function
and in fact replacing it with `smart-space' which has all the logic
in it.