[git] Merge fixup and fix aliases into a single alias.
If called with no arguments, “git fix” will act like “git fixup”
used to, ie. it'll ammend the HEAD commit updating the tree.  If
called with at least one argument, it will act like “git fix” used
to, ie. will create a “fixup!” commit for a commit specified as an
argument.

Also, instead of using a shell function to create the “fixup!”
commit, the alias now uses “git commit --fixup” for that purpos.