[update] Always show diff in interactive mode.
I always ended up asking for the diff to be displayed so to skip
this step, making the script show the diff automatically.

Also, because I've never used reverse diff and now diff option is
not that useful (since it's shown by default) removing those two
from the prompt.

Finally, I've never used “exit”, the letter is confusing as I
always think “edit” when I see it, and terminating script can be
accomplished by either ^C or ^D, so removing that as well.

This ends up with only “y” and “n” as possible answers.