[sawfish] Use “music” script to control music playback.
At times, instead of using MPD, I use Spotify to listen to music.
To be able to control both with the same interface, this commit
creates a “music” script which detects whether Spotify is running
and either issues Spotify or MPD commands.  This commit adjusts
Sawfish configuration as well, to use this new script.