[emacs] Removed unused and redundant code plus small improvements
Mostly removing of code that I've used for a very long while (or
have never used at all) and code that sets variable to its
default value which is redundant.

Beside that some small improvements (e.g. use of `bound-and-true-p`)
and some changes which try to get rid of various compiler warnings.