[emacs] Close read-only buffers with “q”
I often hit “q” in read-only buffers (such as source Elisp files)
treating them a bit like help buffers (and other such buffers), but
that results in Emacs screaming at me.  To fix that, this commit
changes binding of “q” so that it kills current buffer if it's
read-only, and inserts “q” otherwise.