[shell] Add top-level git directory name to git prompt
If inside of a git repository add the name of the top-level directory
next to the branch name, but do that only if there are no other
modifiers (namely there is no merge, rebase, etc. going on) and the
name of the directory is short (i.e. less than 4 letters).

This helps with a situation where the same repository is mirrored
or cloned several times which makes it hard to imposible to
distinguish the working directory once CWD is deep inside of it.