[irssi] format notices the same way messages are formatted
irssi (and basically all IRC clients) have very verbose and weird
formatting of NOTICEs.  There is no apparent reason for that
though, plus it makes output from bots that use NOTICEs rather
ugly. Format notices similarly to the way messages are formatted
except that “>” is replaced with “]” – a very non-intrusive
difference that indicates if something was a notice.