[emacs] Make sure Unicode-capable font is used
I’m not sure why this occurs, but Emacs would pick an ISO-8859-1 version
of the font which would result in ugly Unicode characters.  Make really
sure that Unicode-capable version of the font is used.

Using `set-face-attribute' should be enough as far as I understand, but
it wasn’t so also add a hook to `after-make-frame-functions'.