[sawfish] Add `mn-display-window' function
In my usage of `display-window' it doesn’t always focus window it’s
supposed to.  Problem happens when view-port is changed and cursor is
on top of another window on that new view-port.  In those cases, because
focus follows mouse, the window under cursor ends up focused.

Fix it by replacing `display-window' function with an implementation
which warps cursor to the centre of the (visible portion) of the window
if workspace or viewport change happened.