[sawfish] Add `mn-get-term' function for focusing terminal window
urxvt has a weird bug where when tabbed extension is used some keybindings
do not work properly if mouse cursor is not on top of the window.  Fix that
by warping mouse cursor to focused terminal window.

Furthermore, if terminal window is already focused and there is another one
in current view-port, switch to that one.  Previously, it was impossible to
cycle between two terminal windows with W-$; it can be done now.