[m] Make use of mpc’s pause-maybe command
Implement pause-maybe command using mpc’s pause-if-playing.  This relies on
http://bugs.musicpd.org/view.php?id=4403 not-yet-committed patch (and the
fact that I have a custom-built mpc binary with that patch applied in
~/bin).

With this change, pause-maybe does no longer require two separate
invocations of mpc, so use it in xlocker even when we’re running MPD
(previously it was used only when playing music via Spotify).