[emacs] Remove --preopen-cmd flag from ‘e’ command
I’ve ended up never using the --preopen-cmd flag for the ‘e’ command.  It
was meant to work with Perforce so I can mark file as opened for edit
before opening it in Emacs, but I ended up never using Perforce directly.
Remove that flag and while at it make use of ‘exec’.