[x] Add mapping for superscript, ellipsis and micro to local keymap
Since upgrade to new Debian a lot of key mappings which I was used to
disappeared.  I’m not entirely sure why but my bet is that Debian decided
to patch xkeyboard-config for no reason in a way that broke stuff.
I probably should find the actual cause of the change but it’s much
easier to just add mapping that I want to my local keymap.