[x] Refactor xinitrc; merge it with xinit-kbd
Extend xinitrc by including xinit-kbd and .xscreen-config into it.
This reduces the amount of files laying around my home directory.

While at it, add a fallback running xterm if Sawfish is not installed
on the system.  This is unlikely but may happen on a freshly
configured Slackware installation.

A side effect is now xset is called only once.  Not that it matters
much but I hate being wasteful. ;)