[sh] Simplify ~/.dir-colors handling
Instead of a decision tree in ~/.shellrc file which decides how to best
set LS_COLORS environmnt variable from ~/.dir_colors or some other
file generate shell code for setting LS_COLORS in the update script
and save the result in ~/.dir-colors directly.  With that, ~/.shellrc
can just source the file and be done with it.