[x] Protect against parsing errors when file is edited while running
Since ~/.xinitrc script remains running as long as X session is on
I often end up editing it while shell still interprets it.  This
may results in garbage being executed.  Avoid that by wrapping the
main entry code in a function and having it terminate the script
(with exit or exec).