[x] Make sure temporary file used by .xinitrc is properly deleted
In certain situations¹, the tepmorary file used to by the script to
gather Sawfish’s intent to poweroff or reboot the machine may remain
in place.  Change the script so that it’s never the case.

¹ On exec, shell does not execute trap 0 commands which means that if
exec_quit_cmd executes a quit command the temporary file is never
deleted.  If, in addition, those quit commands fail, the temporary
file is left behind and the machine is not halted or rebooted.

In practice, this should not be an issue since those commands
will reboot or halt the machine thus nuking /tmp, but in case that fails