[x] Map LSGT into a shift
On UK and Swiss keyboard layouts (and probably many more), left shift is
shorter and there’s a less-greater key shoved between shift and ‘z’.  AS a
result, trying to press right edge of what is shift on US layout hits this
additional key.  Map it to shift so I don’t keep making that mistake.