Deduplicate configuration in in Xresources
XTerm and URxvt use the same colours so specify them once for both rather
then once for each.  This has slightly different meaning now since the
settings now apply to all other applications which read ‘color#’ resource
but that might be a good thing.

On the other hand, keep font configuration separate (though make it the
same) in fear of affecting unrelated applications.  More testing is needed
before deduplicating that as well.