[sh] Fix unnecessary escaping of non-English characters in ls output
Setting LANG variable to C clears console encoding information which
causes GNU ls to escape non-ASCII characters in the output.  The only
purpose of overwriting the variable was to ensure English lexical
ordering, but LC_COLLATE is already exported (further down the file)
so there’s no need for any of it.