[shell] Disable git prompt on sshfs mounts as well
sshfs is a remote file system just like nfs so treat it the same way in
the git prompt (that is to say, don’t enable git prompt when on sshfs
file system).

sshfs is uses fuse so just using stat isn’t enough to reliably detect if
a directory is on sshfs (or in some other fuse mount point) but it’s good
enough for now and I likely won’t need git prompt on any other fuse mounts
either.