[shell] Export PAGER and LESS variables
Export PAGER and LESS variables to make sure they are available to executed
commands.  Previously, depending on the system, if the variables weren’t
already in the environment, they would never appear there and e.g. less
options would not be in effect.

While at it, execute less termcap configuration conditionally on less being
available.  In practice, less is available on all systems I care about so
that probably doesn’t matter but it feels right. ;)