[shell] Remove FIGNORE customisation
FIGNORE ends up being more of an inconvenience than what it’s worse.  If
ignored file is the only one that can complete a path, bash will refuse to
acknowledge it because of the FIGNORE.  This forces user to type in the
full name of the file.  Any benefits created by FIGNORE are overshadowed
by this annoyance.