[emacs] Customise mode-line
Use face to indicate whether buffer is modified or not (buffer name is
displayed in italic if it is) which removes one of the asterisk from the
mode-line.

Use » as read-only indicator which I find more pleasing than an asterisk.

Apparently using ‘(format-mode-line "%l")’ wrapped around ‘save-excursion’
is more efficient than counting lines with ‘count-lines’ so do the former
when displaying total number of lines in the buffer.

Finally, remove some other status information I don’t care about and use
subdued colour for some of the punctuation.