[sawfish] Rework aduio volume bindings
Move volume-control keymap from W-x to W-u which is easier to type.
Change the keymap to use 12 keys on the left of the key board to
introduce full volume control form 5% to 100% as well as mute handling.
Lastly, add bindings for audio-related multimedia keys: mute, mute mic,
volume up and volume down.