[emacs] Sort ‘package-selected-packages’ list
Each installed package is added to the ‘package-selected-packages’ list
which then ends up in the custom.el file.  That’s all good but because
the list is unsorted it’s often hard to interpret the difference between
old and new custom.el.  Solve this issue by making sure the variable is
always sorted.