[bash] Put job into background when pressing C-z
Pressing C-z when another process is in the background stops that process.
Pressing it again once back in bash prompt does nothing.  I very often end
up putting jobs into background as soon as I stop them which is a somewhat
tedious ‘C-z bg RET’ sequence to type. Update binding for C-z so that bash
executes ‘bg’ command when it’s pressed.  With this, ‘C-z C-z’ is all that
needs to be typed to stop a process and sand it into background.