[emacs] Cleanup checks of whether features are available
remember-notes, isearch, uniquify and deamon-mode are features which are
always available on any recent Emacs.  No need to wrap their usage in
fboundp-checks, load-checks or add-hook.

Meanwhile, form-feed-mode is a third-party package so check that with
an fboundp.  This fixes a compilation warning when a proper #' prefix
for functions is used.