[x] Store ssh-agent’s socket in known location
Rather than having ssh-agent choose where to put its socket, configure it
such that it puts it in a known path.  The idea is that this will allow
Emacs figure out the path when it needs to make SSH connections.  At the
moment it doesn’t get SSH_AUTH_SOCK variable set because it’s started prior
to X.