[sawfish] Fix switching viewport to browser window
If a viewport is switched while Firefox has input focus and the switch does
not result in the browser loosing focus (e.g. the move was to an empty
viewport), `mn-get-browser` fails to switch viewport back to the one with
the browser.  Ignore focuse if focused window is outside of the current
viewport so that in those scenarios the viewport is moved even if input
focus doesn’t need to be changed.