[sh] update dir_colors
Firstly, do not highlight DOS and Windows executable files based on the
file extension.  I never found it useful while it is sometimes confusing
when I try to localise an executable file for a program with binaries for
multiple platforms.

Secondly, add extensions for AVIF and JPEG XL files to list of images.